+++ src/main/java/com/example/user_management_service/model/dto/FieldWithQuantityDTO.java
@@ -14,6 +14,6 @@ public class FieldWithQuantityDTO {
 
     private Long id;
     private Field fieldName;
-    private Integer quote;
+    private Long quote;
 
 }
